Add vendor-specific initialization hooks.
Allow optional vendor-specific initializations
within init.  This can be used for runtime
initialization setup that init rc scripts do
not support.

Example, in OnePlus 3/T device tree there is "libinit_oneplus3"
which determines whether the device model will be a OP3 or OP3T.
It does not work without this, and as a result the device name ends
up just being "du_oneplus3".

Change-Id: I7623a0d59b18f9ec8e3623958e2f7ccd72b877bf